rename: 메인 게시물 리스트 조회 파일명을 postList로 변경
- 재사용하기위해 범용적 네이밍으로 변경
