+++ src/api/postList/repository/postList.repository.ts
@@ -1,5 +1,5 @@
 import { PrismaClient } from "@prisma/client";
-import { MainPostsResponseDto } from "../dto/mainPostsResponse.dto";
+import { postListResponseDto } from "../dto/postListResponse.dto";
 
 const prisma = new PrismaClient();
 
@@ -7,7 +7,7 @@ export const getPostList = async (
   filters: any,
   page: number,
   limit: number
-): Promise<MainPostsResponseDto[]> => {
+): Promise<postListResponseDto[]> => {
   const skip = (page - 1) * limit;
 
   const posts = await prisma.post.findMany({
